Complete task 14 with several helper methods: action_to_get_enough_resource, state_after_swap, state_after_trade, state_after_building, struc_need_to_built, plan_for_settlement, plan_for_knight, plan_for_city, plan_for_road, indexToSwap.
For task 12, when the current board_state and available resources can support both trade and swap, change from considering actions of trade first to consider swapping first.
That's because each resource has its corresponding unused knight, and swapping for certain types of resource requires the existence of corresponding unused knights in the board_state. However, when trading resources, as long as we have golds, we can use it for trade for any resource. If we perform trading first, it might trade for some resources that having corresponding unused knight in the board_state, and that might make it not able to get enough resource for building target structures.

Add authorship details to task 9-14 and their helper methods